Make the main_test the entrypoint and move senario generation to separate file
